+++ Java/CoreJava/src/main/java/coding/unitTest/AddService.java
@@ -1,4 +1,4 @@
-package org.DsAndAlgo;
+package coding.unitTest;
 
 public interface AddService {
     public int add(int num1, int num2);
