+++ Java/CoreJava/src/main/java/coding/unitTest/AddServiceImpl.java
@@ -1,4 +1,4 @@
-package org.DsAndAlgo;
+package coding.unitTest;
 
 public class AddServiceImpl implements AddService {
     @Override
