+++ Java/CoreJava/src/main/java/coding/unitTest/CalcService.java
@@ -1,4 +1,4 @@
-package org.DsAndAlgo;
+package coding.unitTest;
 
 public class CalcService {
