+++ Java/CoreJava/src/main/java/coding/unitTest/Calculator.java
@@ -1,4 +1,4 @@
-package org.DsAndAlgo;
+package coding.unitTest;
 
 public class Calculator {
