+++ Java/CoreJava/src/main/java/coding/unitTest/StudentDao.java
@@ -1,4 +1,4 @@
-package org.DsAndAlgo;
+package coding.unitTest;
 
 public class StudentDao {
