+++ Java/CoreJava/src/test/java/CalcService2Test.java
@@ -1,21 +1,27 @@
-package org.DsAndAlgo;
-
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
 
-public class CalcService1Test {
+public class CalcService2Test {
+
+    CalcService calcService;
+
+    @Mock
+    AddService addService;
+
+    @BeforeEach
+    public void setup() {
+        MockitoAnnotations.initMocks(this);
+    }
 
     @Test
-    void testCalc() {
+    public void testCalc() {
         System.out.println("**--- Test testCalc executed ---**");
 
-        AddService addService;
-        CalcService calcService;
-
-        addService = Mockito.mock(AddService.class);
         calcService = new CalcService(addService);
 
         int num1 = 11;
