+++ Java/CoreJava/src/test/java/MainTest.java
@@ -1,5 +1,3 @@
-package org.DsAndAlgo;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
