+++ Java/CoreJava/src/test/java/MockSingleElementTest.java
@@ -1,5 +1,3 @@
-package org.DsAndAlgo;
-
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
