+++ Java/CoreJava/src/test/java/ServiceTest.java
@@ -1,6 +1,3 @@
-package org.DsAndAlgo;
-
-
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import static org.junit.jupiter.api.Assertions.*;
