+++ Java/CoreJava/src/test/java/StudentServiceMockitoAnnotationTest.java
@@ -1,5 +1,3 @@
-package org.DsAndAlgo;
-
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
