--- build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'org.DsAndAlgo'
-version = '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-def mockitoversion = '4.0.0'
-
-dependencies {
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-    implementation group: 'org.mockito', name: 'mockito-junit-jupiter', version: '5.10.0'
-    implementation group: 'org.mockito', name: 'mockito-core', version: '5.10.0'
-}
-
-test {
-    useJUnitPlatform()
-}
